requirements.txt
python-dotenv
langchain
langchain-community
langchain_openai
pymupdf
pypdf
faiss-cpu
streamlit
fastapi
uvicorn
python-multipart